--- ark-di/arik-di-android/src/main/java/com/re4rk/arkdiAndroid/arkGenerator/ArkGeneratorExt.kt
@@ -1,7 +1,0 @@
-package com.re4rk.arkdiAndroid.arkGenerator
-
-fun arkGenerator(block: ArkGeneratorBuilder.() -> Unit): ArkGenerator {
-    val builder = ArkGeneratorBuilder()
-    builder.block()
-    return builder.build()
-}
