+++ ark-di/arik-di-android/src/main/java/com/re4rk/arkdiAndroid/arkModules/ArkModules.kt
@@ -1,9 +1,9 @@
-package com.re4rk.arkdiAndroid.arkGenerator
+package com.re4rk.arkdiAndroid.arkModules
 
 import android.content.Context
 import com.re4rk.arkdi.ArkContainer
 
-interface ArkGenerator {
+interface ArkModules {
     fun createApplicationModule(applicationContext: Context): ArkContainer
     fun createRetainedActivityModule(parent: ArkContainer, context: Context): ArkContainer
     fun createActivityModule(parent: ArkContainer, context: Context): ArkContainer
