+++ ark-di/arik-di-android/src/main/java/com/re4rk/arkdiAndroid/arkModules/ArkModulesExt.kt
@@ -1,0 +1,7 @@
+package com.re4rk.arkdiAndroid.arkModules
+
+fun arkModules(block: ArkModulesBuilder.() -> Unit): ArkModules {
+    val builder = ArkModulesBuilder()
+    builder.block()
+    return builder.build()
+}
